# docs/swagger/swagger.yaml
basePath: /core/v1
definitions:
...
  lakukan-be_internal_modules_core_permission_template_dto.CreatePermissionTemplateRequest:
    properties:
      actions:
        items:
          type: string
        minItems: 1
        type: array
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - actions
    - name
    type: object
  lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      templates:
        items:
          $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateResponse:
    properties:
      actions:
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      deleted_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  lakukan-be_internal_modules_core_permission_template_dto.UpdatePermissionTemplateRequest:
    properties:
      actions:
        items:
          type: string
        minItems: 1
        type: array
      description:
        maxLength: 1000
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - actions
    - name
    type: object
...
  lakukan-be_internal_shared_response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Pagination'
    type: object
  lakukan-be_internal_shared_response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  lakukan-be_internal_shared_response.Response:
    properties:
      data: {}
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
      meta:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@lakukan.com
    name: API Support
    url: http://www.lakukan.com/support
  description: Backend API untuk Lakukan - Multi-tenant Business Management Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lakukan API
  version: "1.0"
paths:
...
  /permission-templates:
    get:
      consumes:
      - application/json
      description: Get list of permission templates with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by name or description
        in: query
        name: search
        type: string
      - description: Include system templates
        in: query
        name: include_system
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get all permission templates
      tags:
      - Permission Templates
    post:
      consumes:
      - application/json
      description: Create a new permission template with actions
      parameters:
      - description: Create permission template request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.CreatePermissionTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create a new permission template
      tags:
      - Permission Templates
  /permission-templates/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a permission template
      parameters:
      - description: Permission Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a permission template
      tags:
      - Permission Templates
    get:
      consumes:
      - application/json
      description: Get a single permission template with its actions
      parameters:
      - description: Permission Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get permission template by ID
      tags:
      - Permission Templates
    put:
      consumes:
      - application/json
      description: Update an existing permission template
      parameters:
      - description: Permission Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Update permission template request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.UpdatePermissionTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/lakukan-be_internal_modules_core_permission_template_dto.PermissionTemplateResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update a permission template
      tags:
      - Permission Templates
  /permission-templates/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft deleted permission template
      parameters:
      - description: Permission Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Restore a permission template
      tags:
      - Permission Templates
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"